Potential fix for code scanning alert no. 4: Multiplication result converted to larger type

+++ src/hw/pvscsi.c
@@ -243,7 +243,7 @@ pvscsi_process_op(struct disk_op_s *op)
     req->tag = SIMPLE_QUEUE_TAG;
     req->flags = scsi_is_read(op) ?
         PVSCSI_FLAG_CMD_DIR_TOHOST : PVSCSI_FLAG_CMD_DIR_TODEVICE;
-    req->dataLen = op->count * blocksize;
+    req->dataLen = (u64)op->count * (u64)blocksize;
     req->dataAddr = (u32)op->buf_fl;
     s->reqProdIdx = s->reqProdIdx + 1;
